about to change dockerfile

+++ NoteFinder.Interfaces/NoteFinderInterfaces.cs
@@ -35,7 +35,7 @@
             set;
         }
 
-        int ChromaticPosition
+        byte ChromaticPosition
         {
             get;
             set;
